Templates folder for generator files. Moved generate config to separate module.
