+++ gulpfile.js
@@ -2,53 +2,17 @@ require('dotenv').config();
 
 const fs = require('fs');
 const gulp = require('gulp');
-const replace = require('gulp-replace');
 const run = require('gulp-run-command').default;
-const propertiesPath = './src/common';
+const generateConfig = require('./scripts/gulp-generate-config');
+
 const pwaAssetGenerator = require('pwa-asset-generator');
 const consumerPublicPath = './src/consumer/public';
 
-
 /**
- * 
- * `gulp properties`
- * 
- * Save the properties module output as json 
- * for consumption by server and client
- * 
- * JSON files generated should not be committed to the codebase
- * Modify generate/*.properties.js to add more properties
- * 
+ * `gulp generateConfig`
+ * generates properties files and sitemap config for consumer
  */
-
-const commonProps = require('./scripts/common.properties');
-
-// generate common.properties.json for client and server consumption
-const generateCommonProperties = (done) => {
-    const filePath = `${propertiesPath}/common.properties.json`;
-    const value = JSON.stringify(commonProps, null, 2);
-    try {
-        fs.writeFile(filePath, value, done);
-    } catch (error) {
-        done(error);
-    }
-}
-
-// generate next-sitemap config with siteUrl value
-const generateNextSiteMapConfig = async (done) => {
-    try {
-        await gulp.src('./generate/next-sitemap.js')
-            .pipe(replace('/*URL*/', commonProps.APPLICATION_URL))
-            .pipe(gulp.dest('./'));
-        done();
-    } catch (error) {
-        done(error);
-    }
-}
-
-const config = gulp.series(generateCommonProperties, generateNextSiteMapConfig)
-
-exports.config = config;
+exports.generateConfig = generateConfig;
 
 /**
  *
@@ -57,7 +21,6 @@ exports.config = config;
  * Run local Next.js development server
  *
  */
-
 const dev = async (done) => {
     try {
         await run('next src/consumer/')();
@@ -100,7 +63,7 @@ exports.buildConsumer = buildConsumer;
 const brandPath = `${consumerPublicPath}/brand`;
 
 // generate common.properties.json for client and server consumption
-const manifest = require('./scripts/manifest');
+const manifest = require('./scripts/templates/manifest');
 const generateManifest = (done) => {
     const filePath = `${consumerPublicPath}/manifest.json`;
     const value = JSON.stringify(manifest, null, 2);
@@ -123,7 +86,7 @@ const clearAssets = async (done) => {
 const generateFavicon = async (done) => {
     try {
         await pwaAssetGenerator.generateImages(
-            './gravit/icon.svg', //TODO: env
+            './vector/icon.svg', //TODO: env
             `${brandPath}/favicon`,
             {
                 scrape: true,
@@ -142,7 +105,7 @@ const generateFavicon = async (done) => {
 const generateSplashScreen = async (done) => {
     try {
         await pwaAssetGenerator.generateImages(
-            './gravit/logo.svg', //TODO: env
+            './vector/logo.svg', //TODO: env
             `${brandPath}/splash`,
             {
                 scrape: true,
