+++ scripts/templates/common.properties.js
@@ -11,7 +11,10 @@ const properties = {
     APPLICATION_URL: process.env.APPLICATION_URL,
     APPLICATION_NAME: process.env.APPLICATION_NAME,
     APPLICATION_SHORT_NAME: process.env.APPLICATION_SHORT_NAME,
-    APPLICATION_DESCRIPTION: process.env.APPLICATION_DESCRIPTION
+    APPLICATION_DESCRIPTION: process.env.APPLICATION_DESCRIPTION,
+
+    // C
+    CONSUMER_URL: process.env.CONSUMER_URL,
 
 };
