+++ scripts/templates/manifest.js
