+++ scripts/templates/next-sitemap.js
@@ -1,5 +1,5 @@
 module.exports = {
-    siteUrl: '/*URL*/',
+    siteUrl: '/*CONSUMER_URL*/',
     generateRobotsTxt: true,
     sourceDir: './dist/consumer', //TODO: env
     outDir: './src/consumer/public' //TODO: env
